CA-107425: Adding setup file for installing just the acktools.
This is a work around for the fact that we are currently not repackaging the python modules to use the distutils install mechanism. We do however want to be able to share some code between both the plugin and the kit. This allows us to install a library that can be used in such a way, and avoid duplication of effort.

There are a number of function calls in utils/autocertkit which we would like to move into this package over time.

Signed-off-by: Rob Dobson <rob@rdobson.co.uk>