CA-147967: Don't convert the reboot flag's timestamp to GMT.
Keep the local time of the reboot flag in order to compare with the crashdump time handed back by XAPI. We might even want to think writing the timestamp out-to-file
to ensure that the local time doesn't become out of sync with XAPI's time. This should however not be the case. Also added some unittests for reading/writing flags.

Swapping to use time.strptime instead of datetime.strptime due it's
non-existence in python 2.4 (as pointed out by Jun).

Signed-off-by: Rob Dobson <rob.dobson@citrix.com>
Reviewed-by: Jun Wook Kwak <junwook.kwak@citrix.com>