CA-148382: Make sure the auto-cert-kit script doesn't get deactivated on upgrade
Unfortunatly, due to the way RPM scriplets get executed, the chkconfig --del
auto-cert-kit command was being executed on upgrade after running the new
packages install scripts. This stopped re-running after reboot from working.

This patch should address that issue and make sure we only remove the chkconfig
entry when the RPM is actually being removed (in which case the next RPM to be
installed will setup the chkconfig entry correctly).

Signed-off-by: Rob Dobson <rob.dobson@citrix.com>