Change deploy_two_vms to use synchronous call to start VMs.
When 2 VMs are created and starts on each host, async call may not
return results. (CA-146164) Even if we ignore response of async
calls, XAPI may fail to track VMs' power and poeration status.
(CA-160978) We can try force shutdown VMs, but it sometimes halts
during the operation. (CA-161590) To avoid these XAPI issues,
deploy_two_vms will starts VMs synchronously until issues are fixed.

As this function is not being used in operation tests (which means
all tests that use this function are not testing XS functionality
such as async call or plugin call), it won't affect result of tests
or  certification.