CP-46833: Fix 'NoneType' object is not subscriptable
If get_dev_pci_addr() does not match the __pci_addr_regex in the
symlink in /sys/dev/block/* that matches the disk, it returns (-1,
None). Nothing in plugins/autocertkit:840 checks for the error and
_is_bus_id() blindly indexes its first argument directly: bus_id[-7:].
If bus_id was None, this would have raised an Exception.

Make _is_bus_id() more robust by checking that bus_id is not None
before attempting to subscript it.